database: modularized database package methods and structs into separated files for more readability
