+++ src/database/get_db.go
@@ -1,0 +1,9 @@
+package database
+
+func GetDB() DB {
+	var db DB
+
+	db.Init()
+
+	return db
+}
