+++ src/database/init_db.go
@@ -1,0 +1,9 @@
+package database
+
+func (db *DB) Init() {
+	db.Client = createClient()
+	println("Client initialized succesfully ✅")
+
+	db.Collections = getAllCollections(db.Client)
+	println("Collections initialized succesfully ✅")
+}
